DSDA: Bypass AOSP DSDA logic of account selection
Internal DSDA solution is failed due to new DSDA logic of
account selection is enabled with feature flag
simultaneous_calling_indications with that we do not allow a
new outgoing call on another PhoneAccount if the active
PhoneAccount can not support a call on a different account.
Fixing by disabling the new AOSP logic by using existing API
isDsdaOrDsdsTransitionMode().

Change-Id: Id404b97955808f174f1bae3f22a87ee11d9f9c7c
CRs-Fixed: 3798977