IMS: Fix CTS failure when DSDS transition is supported
Some CTS tests try to initiate call with target phone account
handle not matching the available registered phone accounts.
The tests expect the call to be initiated on an available
phone account handle. Current check for DSDS transition mode
returns empty list of accounts when the phone account handles
are not matching. Restrict this check to only telephony calls
to prevent the CTS failures.

CRs-Fixed: 3825888
Change-Id: I7bdc4ff024df338c766c72b663fb4bf6af5c97c2